completed algorithm

+++ functions/add/add.h
@@ -28,7 +28,7 @@ class add: public primitive
             set(1, right);
         }
 
-        float eval(Features& features)
+        double eval(Features& features)
         {
             return get(0)->eval(features) + get(1)->eval(features);
         }
