+++ terminal/ground_temperature/ground_temperature.h
@@ -19,7 +19,7 @@ class ground_temperature: public primitive
             return result;
         }
 
-        float eval(Features& features)
+        double eval(Features& features)
         {
             return features.ground_temperature;
         }
