+++ terminal/temperature/temperature.h
@@ -19,7 +19,7 @@ class temperature: public primitive
             return result;
         }
 
-        float eval(Features& features)
+        double eval(Features& features)
         {
             return features.temperature;
         }
